backend/requirements.txt
fastapi
uvicorn
praw
transformers
pydantic
httpx
python-dotenv 